Making it work with Rails 2.3 and Ruby 1.8
I was explicitly requiring ActiveRecord and ActiveSupport 3.2.x, but I've dropped the version numbers (there's no reason, I'm not using anything exclusive to a version of AR or AS).
I was also using the new Ruby 1.9 hash syntax that looks like JSON syntax, but it wasn't working with Ruby 1.8, so I switched.