Adding session management; re-arranging code
First off, I made everything have a top level module name of GoogleAuthenticatorRails to match the name of the gem.  Now we're not polluting ActiveRecord or Google's namespace.

Secondly, I added internal session management in the style of Authlogic.  This utilizies a persistence_token (that is assumed available on the model) to save some data to a cookie to notify that the user is authenticated.