Address `Bundler::GemfileError` error
When running the appraisal tests, the following error was occurring:
```
ArgumentError: Trying to register Bundler::GemfileError for status code 4 but Bundler::GemfileError is already registered
```

This appears to be related to an error introduced either in Bundler 2.1+
or in Rails 5.2+ (or in rubgems < 3.1), which I've not been able to
exactly track down.

When specifying the version of bundler for the CircleCI ruby orb to
install, as this commit does, the `install-deps` command explicitly
installs the bundler gem, then runs `bundle install` which notices that
the version of bundler specified in the lockfile is later and installs
that instead. The upshot is that we're still using the version of
bundler specified in the lockfile to run the tests, but for some reason
installing a "working" bundler gemfile version first means we don't then
see the original `Bundler::GemfileError`.

In short, I don't understand why this fixes the problem